Update admin.html
missed to close the DIV container in line 173, and that made me crazy because after logging in as admin, it wasn't showing anything. I thought there was something wrong with the app script code or the connection is wrong. I tried and changed everything and wasted 3 hours and then somehow I got to know your code editor may not show you an error for not closing <div> but that can totally mess up your page or show nothing at all. I am installing HTMLHint extension in VS code now to prevent this kind of mistakes. Lesson Learned.